pyproject.toml
[tool.poetry]
name = "testtask"
version = "0.1.0"
description = ""
authors = ["Yevgeniy Kap <kapustaevg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
setuptools = "^69.0.3"
stripe = "^7.10.0"
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
